Don't pull images on every start
This changes the default of the "--pull" argument for the "start"
command to false. It is not necessary to (force-) pull images on
every container start and takes a long time to complete. Users may
also run into trouble with upcoming rate limits introduced for Docker
Hub.

Therefore, if you need a new image version, for the time being, either
pull the image manually or use the --pull argument.